PowerPulse/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "powerpulse"
version = "0.1.0"
description = "A real-time OS power consumption analysis tool."
authors = [
    { name = "User", email = "user@example.com" }
]
dependencies = [
    "psutil",
    "rich",
    "fastapi",
    "uvicorn",
    "websockets",
    "google-genai",
    "python-dotenv"
]
requires-python = ">=3.8"

[project.scripts]
powerpulse = "powerpulse.cli:main"
